+++ features/expenses/src/main/kotlin/es/pedrazamiguez/splittrip/features/expense/presentation/model/PaymentBadgeData.kt
@@ -1,0 +1,7 @@
+package es.pedrazamiguez.splittrip.features.expense.presentation.model
+
+data class PaymentBadgeData(
+    val text: String,
+    val isPassed: Boolean,
+    val isToday: Boolean
+)
